Adjust data for the effect of other variable(s)

This function can be used to adjust the data for the effect of other variables present in the dataset. It is based on an underlying fitting of regressions models, allowing for quite some flexibility, such as including factors as random effects in mixed models (multilevel partialization), continuous variables as smooth terms in general additive models (non-linear partialization) and/or fitting these models under a Bayesian framework. The values returned by this function are the residuals of the regression models. Note that a regular correlation between two "adjusted" variables is equivalent to the partial correlation between them.

adjust(
  data,
  effect = NULL,
  select = NULL,
  exclude = NULL,
  multilevel = FALSE,
  additive = FALSE,
  bayesian = FALSE
)

Arguments

data
A dataframe.

effect
Character vector of column names to be adjusted for (regressed out). If NULL (the default), all variables will be selected.

select
Character vector of column names. If NULL (the default), all variables will be selected.

exclude
Character vector of column names to be excluded from selection.

multilevel
If TRUE, the factors are included as random factors. Else, if FALSE (default), they are included as fixed effects in the simple regression model.

additive
If TRUE, continuous variables as included as smooth terms in additive models. The goal is to regress-out potential non-linear effects.

bayesian
If TRUE, the models are fitted under the Bayesian framework using rstanarm.

Examples

adjust(iris, effect = "Species", select = "Sepal.Length")

#>     Sepal.Length Sepal.Width Petal.Length Petal.Width    Species
#> 1          0.094         3.5          1.4         0.2     setosa
#> 2         -0.106         3.0          1.4         0.2     setosa
#> 3         -0.306         3.2          1.3         0.2     setosa
#> 4         -0.406         3.1          1.5         0.2     setosa
#> 5         -0.006         3.6          1.4         0.2     setosa
#> 6          0.394         3.9          1.7         0.4     setosa
#> 7         -0.406         3.4          1.4         0.3     setosa
#> 8         -0.006         3.4          1.5         0.2     setosa
#> 9         -0.606         2.9          1.4         0.2     setosa
#> 10        -0.106         3.1          1.5         0.1     setosa
#> 11         0.394         3.7          1.5         0.2     setosa
#> 12        -0.206         3.4          1.6         0.2     setosa
#> 13        -0.206         3.0          1.4         0.1     setosa
#> 14        -0.706         3.0          1.1         0.1     setosa
#> 15         0.794         4.0          1.2         0.2     setosa
#> 16         0.694         4.4          1.5         0.4     setosa
#> 17         0.394         3.9          1.3         0.4     setosa
#> 18         0.094         3.5          1.4         0.3     setosa
#> 19         0.694         3.8          1.7         0.3     setosa
#> 20         0.094         3.8          1.5         0.3     setosa
#> 21         0.394         3.4          1.7         0.2     setosa
#> 22         0.094         3.7          1.5         0.4     setosa
#> 23        -0.406         3.6          1.0         0.2     setosa
#> 24         0.094         3.3          1.7         0.5     setosa
#> 25        -0.206         3.4          1.9         0.2     setosa
#> 26        -0.006         3.0          1.6         0.2     setosa
#> 27        -0.006         3.4          1.6         0.4     setosa
#> 28         0.194         3.5          1.5         0.2     setosa
#> 29         0.194         3.4          1.4         0.2     setosa
#> 30        -0.306         3.2          1.6         0.2     setosa
#> 31        -0.206         3.1          1.6         0.2     setosa
#> 32         0.394         3.4          1.5         0.4     setosa
#> 33         0.194         4.1          1.5         0.1     setosa
#> 34         0.494         4.2          1.4         0.2     setosa
#> 35        -0.106         3.1          1.5         0.2     setosa
#> 36        -0.006         3.2          1.2         0.2     setosa
#> 37         0.494         3.5          1.3         0.2     setosa
#> 38        -0.106         3.6          1.4         0.1     setosa
#> 39        -0.606         3.0          1.3         0.2     setosa
#> 40         0.094         3.4          1.5         0.2     setosa
#> 41        -0.006         3.5          1.3         0.3     setosa
#> 42        -0.506         2.3          1.3         0.3     setosa
#> 43        -0.606         3.2          1.3         0.2     setosa
#> 44        -0.006         3.5          1.6         0.6     setosa
#> 45         0.094         3.8          1.9         0.4     setosa
#> 46        -0.206         3.0          1.4         0.3     setosa
#> 47         0.094         3.8          1.6         0.2     setosa
#> 48        -0.406         3.2          1.4         0.2     setosa
#> 49         0.294         3.7          1.5         0.2     setosa
#> 50        -0.006         3.3          1.4         0.2     setosa
#> 51         1.064         3.2          4.7         1.4 versicolor
#> 52         0.464         3.2          4.5         1.5 versicolor
#> 53         0.964         3.1          4.9         1.5 versicolor
#> 54        -0.436         2.3          4.0         1.3 versicolor
#> 55         0.564         2.8          4.6         1.5 versicolor
#> 56        -0.236         2.8          4.5         1.3 versicolor
#> 57         0.364         3.3          4.7         1.6 versicolor
#> 58        -1.036         2.4          3.3         1.0 versicolor
#> 59         0.664         2.9          4.6         1.3 versicolor
#> 60        -0.736         2.7          3.9         1.4 versicolor
#> 61        -0.936         2.0          3.5         1.0 versicolor
#> 62        -0.036         3.0          4.2         1.5 versicolor
#> 63         0.064         2.2          4.0         1.0 versicolor
#> 64         0.164         2.9          4.7         1.4 versicolor
#> 65        -0.336         2.9          3.6         1.3 versicolor
#> 66         0.764         3.1          4.4         1.4 versicolor
#> 67        -0.336         3.0          4.5         1.5 versicolor
#> 68        -0.136         2.7          4.1         1.0 versicolor
#> 69         0.264         2.2          4.5         1.5 versicolor
#> 70        -0.336         2.5          3.9         1.1 versicolor
#> 71        -0.036         3.2          4.8         1.8 versicolor
#> 72         0.164         2.8          4.0         1.3 versicolor
#> 73         0.364         2.5          4.9         1.5 versicolor
#> 74         0.164         2.8          4.7         1.2 versicolor
#> 75         0.464         2.9          4.3         1.3 versicolor
#> 76         0.664         3.0          4.4         1.4 versicolor
#> 77         0.864         2.8          4.8         1.4 versicolor
#> 78         0.764         3.0          5.0         1.7 versicolor
#> 79         0.064         2.9          4.5         1.5 versicolor
#> 80        -0.236         2.6          3.5         1.0 versicolor
#> 81        -0.436         2.4          3.8         1.1 versicolor
#> 82        -0.436         2.4          3.7         1.0 versicolor
#> 83        -0.136         2.7          3.9         1.2 versicolor
#> 84         0.064         2.7          5.1         1.6 versicolor
#> 85        -0.536         3.0          4.5         1.5 versicolor
#> 86         0.064         3.4          4.5         1.6 versicolor
#> 87         0.764         3.1          4.7         1.5 versicolor
#> 88         0.364         2.3          4.4         1.3 versicolor
#> 89        -0.336         3.0          4.1         1.3 versicolor
#> 90        -0.436         2.5          4.0         1.3 versicolor
#> 91        -0.436         2.6          4.4         1.2 versicolor
#> 92         0.164         3.0          4.6         1.4 versicolor
#> 93        -0.136         2.6          4.0         1.2 versicolor
#> 94        -0.936         2.3          3.3         1.0 versicolor
#> 95        -0.336         2.7          4.2         1.3 versicolor
#> 96        -0.236         3.0          4.2         1.2 versicolor
#> 97        -0.236         2.9          4.2         1.3 versicolor
#> 98         0.264         2.9          4.3         1.3 versicolor
#> 99        -0.836         2.5          3.0         1.1 versicolor
#> 100       -0.236         2.8          4.1         1.3 versicolor
#> 101       -0.288         3.3          6.0         2.5  virginica
#> 102       -0.788         2.7          5.1         1.9  virginica
#> 103        0.512         3.0          5.9         2.1  virginica
#> 104       -0.288         2.9          5.6         1.8  virginica
#> 105       -0.088         3.0          5.8         2.2  virginica
#> 106        1.012         3.0          6.6         2.1  virginica
#> 107       -1.688         2.5          4.5         1.7  virginica
#> 108        0.712         2.9          6.3         1.8  virginica
#> 109        0.112         2.5          5.8         1.8  virginica
#> 110        0.612         3.6          6.1         2.5  virginica
#> 111       -0.088         3.2          5.1         2.0  virginica
#> 112       -0.188         2.7          5.3         1.9  virginica
#> 113        0.212         3.0          5.5         2.1  virginica
#> 114       -0.888         2.5          5.0         2.0  virginica
#> 115       -0.788         2.8          5.1         2.4  virginica
#> 116       -0.188         3.2          5.3         2.3  virginica
#> 117       -0.088         3.0          5.5         1.8  virginica
#> 118        1.112         3.8          6.7         2.2  virginica
#> 119        1.112         2.6          6.9         2.3  virginica
#> 120       -0.588         2.2          5.0         1.5  virginica
#> 121        0.312         3.2          5.7         2.3  virginica
#> 122       -0.988         2.8          4.9         2.0  virginica
#> 123        1.112         2.8          6.7         2.0  virginica
#> 124       -0.288         2.7          4.9         1.8  virginica
#> 125        0.112         3.3          5.7         2.1  virginica
#> 126        0.612         3.2          6.0         1.8  virginica
#> 127       -0.388         2.8          4.8         1.8  virginica
#> 128       -0.488         3.0          4.9         1.8  virginica
#> 129       -0.188         2.8          5.6         2.1  virginica
#> 130        0.612         3.0          5.8         1.6  virginica
#> 131        0.812         2.8          6.1         1.9  virginica
#> 132        1.312         3.8          6.4         2.0  virginica
#> 133       -0.188         2.8          5.6         2.2  virginica
#> 134       -0.288         2.8          5.1         1.5  virginica
#> 135       -0.488         2.6          5.6         1.4  virginica
#> 136        1.112         3.0          6.1         2.3  virginica
#> 137       -0.288         3.4          5.6         2.4  virginica
#> 138       -0.188         3.1          5.5         1.8  virginica
#> 139       -0.588         3.0          4.8         1.8  virginica
#> 140        0.312         3.1          5.4         2.1  virginica
#> 141        0.112         3.1          5.6         2.4  virginica
#> 142        0.312         3.1          5.1         2.3  virginica
#> 143       -0.788         2.7          5.1         1.9  virginica
#> 144        0.212         3.2          5.9         2.3  virginica
#> 145        0.112         3.3          5.7         2.5  virginica
#> 146        0.112         3.0          5.2         2.3  virginica
#> 147       -0.288         2.5          5.0         1.9  virginica
#> 148       -0.088         3.0          5.2         2.0  virginica
#> 149       -0.388         3.4          5.4         2.3  virginica
#> 150       -0.688         3.0          5.1         1.8  virginica

# \donttest{
if (FALSE) {
adjust(iris, effect = "Species", select = "Sepal.Length", multilevel = TRUE)
adjust(iris, effect = "Species", select = "Sepal.Length", bayesian = TRUE)
adjust(iris, effect = "Petal.Width", select = "Sepal.Length", additive = TRUE)
adjust(iris,
  effect = "Petal.Width", select = "Sepal.Length",
  additive = TRUE, bayesian = TRUE
)
adjust(iris,
  effect = c("Petal.Width", "Species"), select = "Sepal.Length",
  multilevel = TRUE, additive = TRUE
)
adjust(iris)
}
# }