Rescale a numeric variable

Rescale a numeric variable to a new range.

change_scale(x, ...)

# S3 method for numeric
change_scale(x, to = c(0, 100), range = NULL, verbose = TRUE, ...)

# S3 method for grouped_df
change_scale(
  x,
  select = NULL,
  exclude = NULL,
  to = c(0, 100),
  range = NULL,
  ...
)

# S3 method for data.frame
change_scale(
  x,
  select = NULL,
  exclude = NULL,
  to = c(0, 100),
  range = NULL,
  ...
)

Arguments

x
Object.

...
Arguments passed to or from other methods.

to
New range of values of the data after rescaling.

range
Initial (old) range of values. If NULL, will take the range of data.

verbose
Toggle warnings on or off.

select
Character vector of column names. If NULL (the default), all variables will be selected.

exclude
Character vector of column names to be excluded from selection.

Value

A rescaled object.

See also

Examples

change_scale(c(0, 1, 5, -5, -2))

#> [1]  50  60 100   0  30

change_scale(c(0, 1, 5, -5, -2), to = c(-5, 5))

#> [1]  0  1  5 -5 -2

head(change_scale(iris))

#>   Sepal.Length Sepal.Width Petal.Length Petal.Width Species
#> 1    22.222222    62.50000     6.779661    4.166667  setosa
#> 2    16.666667    41.66667     6.779661    4.166667  setosa
#> 3    11.111111    50.00000     5.084746    4.166667  setosa
#> 4     8.333333    45.83333     8.474576    4.166667  setosa
#> 5    19.444444    66.66667     6.779661    4.166667  setosa
#> 6    30.555556    79.16667    11.864407   12.500000  setosa