Effect size for differences

Compute different indices of effect size. For very small sample sizes (n < 20) Hedges' g is considered as less biased than Cohen's d. For sample sizes > 20, the results for both statistics are roughly equivalent.

The Glass’s delta is appropriate if standard deviations are significantly different between groups, as it uses only the second group's standard deviation.

cohens_d(
  x,
  y = NULL,
  data = NULL,
  correction = FALSE,
  pooled_sd = TRUE,
  paired = FALSE,
  ci = 0.95
)

hedges_g(
  x,
  y = NULL,
  data = NULL,
  correction = FALSE,
  pooled_sd = TRUE,
  paired = FALSE,
  ci = 0.95
)

glass_delta(x, y = NULL, data = NULL, correction = FALSE, ci = 0.95)

Arguments

x
A formula, a numeric vector, or a name of one in data.

y
A numeric vector, a grouping (character / factor) vector, a or a name of one in data. Ignored if x is a formula.

data
An optional data frame containing the variables.

correction
If TRUE, applies a correction to make it less biased for small samples (McGrath & Meyer, 2006).

pooled_sd
If TRUE (default), a sd_pooled is used (assuming equal variance). Else the mean SD from both groups is used instead.

paired
If TRUE, the values of x and y are considered as paired.

ci
Confidence Interval (CI) level

Value

A data frame with the effect size(s) and confidence interval(s).

Confidence Intervals

Confidence intervals are estimated using the Noncentrality parameter method; These methods searches for a the best ncp (non-central parameters) for of the noncentral t distribution for the desired tail-probabilities, and then convert these ncps to the corresponding effect sizes.

References

Cohen, J. (2013). Statistical power analysis for the behavioral sciences. Routledge.

McGrath, R. E., & Meyer, G. J. (2006). When effect sizes disagree: the case of r and d. Psychological methods, 11(4), 386.

Hedges, L. V. & Olkin, I. (1985). Statistical methods for meta-analysis. Orlando, FL: Academic Press.

Examples

cohens_d(iris$Sepal.Length, iris$Sepal.Width)

#> Cohen's d |       95% CI
#> ------------------------
#>      4.21 | [3.80, 4.61]

hedges_g("Sepal.Length", "Sepal.Width", data = iris)

#> Hedge's g |       95% CI
#> ------------------------
#>      4.20 | [3.79, 4.60]

cohens_d(mpg ~ am, data = mtcars)

#> Cohen's d |         95% CI
#> --------------------------
#>     -1.48 | [-2.27, -0.67]

cohens_d(mpg ~ am, data = mtcars, pooled_sd = FALSE)

#> Cohen's d |         95% CI
#> --------------------------
#>     -1.41 | [-2.13, -0.56]

hedges_g(mpg ~ am, data = mtcars)

#> Hedge's g |         95% CI
#> --------------------------
#>     -1.44 | [-2.21, -0.65]

glass_delta(mpg ~ am, data = mtcars)

#> Glass' delta |         95% CI
#> -----------------------------
#>        -1.17 | [-1.93, -0.40]