effectsize-package {effectsize}	R Documentation

effectsize: Indices of Effect Size

Description

In both theoretical and applied research, it is often of interest to assess the strength of an observed association. This is typically done to allow the judgment of the magnitude of an effect, especially when units of measurement are not meaningful. Though some indices of effect size, such as the correlation coefficient (itself a standardized covariance coefficient) are readily available, other measures are often harder to obtain.

effectsize fills this important gap, providing utilities for easily estimating a wide variety of standardized effect sizes (i.e., effect sizes that are not tied to the units of measurement of the variables of interest) and their confidence intervals (CIs), from a variety of statistical models and hypothesis tests, such as cohens_d(), phi(), eta_squared(), and many more.

See vignette("effectsize", package = "effectsize") for more details, or vignette(package = "effectsize") for a full list of vignettes.

References: Ben-Shachar et al. (2020) doi: 10.21105/joss.02815.

Details

effectsize

Author(s)

Maintainer: Mattan S. Ben-Shachar mattansb@msbstats.info (ORCID) (@mattansb)

Authors:

Dominique Makowski dom.makowski@gmail.com (ORCID) (@Dom_Makowski)

Daniel Lüdecke d.luedecke@uke.de (ORCID) (@strengejacke)

Indrajeet Patil patilindrajeet.science@gmail.com (ORCID) (@patilindrajeets)

Brenton M. Wiernik brenton@wiernik.org (ORCID) (@bmwiernik)

Rémi Thériault remi.theriault@mail.mcgill.ca (ORCID) (@rempsyc)

Philip Waggoner philip.waggoner@gmail.com (ORCID) [contributor]

Other contributors:

Ken Kelley [contributor]

David Stanley [contributor]

Aaron Caldwell arcaldwell49@gmail.com (ORCID) [contributor]

Jessica Burnett jburnett@usgs.gov (ORCID) [reviewer]

Johannes Karreth jkarreth@ursinus.edu (ORCID) [reviewer]

See Also

Useful links:

https://easystats.github.io/effectsize/

Report bugs at https://github.com/easystats/effectsize/issues/

[Package effectsize version 0.8.7 Index]