Effect size for ANOVA

Functions to compute effect size measures for ANOVAs, such as Eta, Omega and Epsilon squared, and Cohen's f (or their partialled versions) for aov, aovlist and anova models. These indices represent an estimate of how much variance in the response variables is accounted for by the explanatory variable(s).

Effect sizes are computed using the sums of squares obtained from anova(model) which might not always be appropriate (Yeah... ANOVAs are hard...). See details.

eta_squared(model, partial = TRUE, ci = 0.9, ...)

omega_squared(model, partial = TRUE, ci = 0.9, ...)

epsilon_squared(model, partial = TRUE, ci = 0.9, ...)

cohens_f(model, partial = TRUE, ci = 0.9, ...)

Arguments

model
A model, ANOVA object, or the result of parameters::model_parameters.

partial
If TRUE, return partial indices.

ci
Confidence Interval (CI) level

...
Arguments passed to or from other methods (ignored).

Value

A data frame with the effect size(s) and confidence interval(s).

A data frame containing the effect size values and their confidence intervals.

Details

For aov and aovlist models, the effect sizes are computed directly with Sums-of-Squares. For all other model, the model is passed to anova(), and effect sizes are approximated via test statistic conversion (see F_to_eta2 for more details.)

Type of Sums of Squares

The sums of squares (or F statistics) used for the computation of the effect sizes is based on those returned by anova(model) (whatever those may be - for aov and aovlist these are type-1 sums of squares; for merMod these are type-3 sums of squares). Make sure these are the sums of squares you are intrested in (you might want to pass the result of car::Anova(mode, type = 3)).

It is generally recommended to fit models with contr.sum factor weights and centered covariates, for sensible results. See examples.

Confidence Intervals

Confidence intervals are estimated using the Noncentrality parameter method; These methods searches for a the best ncp (non-central parameters) for of the noncentral F distribution for the desired tail-probabilities, and then convert these ncps to the corresponding effect sizes.

Omega Squared

Omega squared is considered as a lesser biased alternative to eta-squared, especially when sample sizes are small (Albers \& Lakens, 2018). Field (2013) suggests the following interpretation heuristics:

Omega Squared = 0 - 0.01: Very small

Omega Squared = 0.01 - 0.06: Small

Omega Squared = 0.06 - 0.14: Medium

Omega Squared > 0.14: Large

Epsilon Squared

It is one of the least common measures of effect sizes: omega squared and eta squared are used more frequently. Although having a different name and a formula in appearance different, this index is equivalent to the adjusted R2 (Allen, 2017, p. 382).

Cohen's f

Cohen's f can take on values between zero, when the population means are all equal, and an indefinitely large number as standard deviation of means increases relative to the average standard deviation within each group. Cohen has suggested that the values of 0.10, 0.25, and 0.40 represent small, medium, and large effect sizes, respectively.

References

Albers, C., \& Lakens, D. (2018). When power analyses based on pilot data are biased: Inaccurate effect size estimators and follow-up bias. Journal of experimental social psychology, 74, 187-195.

Allen, R. (2017). Statistics and Experimental Design for Psychologists: A Model Comparison Approach. World Scientific Publishing Company.

Field, A. (2013). Discovering statistics using IBM SPSS statistics. sage.

Kelley, K. (2007). Methods for the behavioral, educational, and social sciences: An R package. Behavior Research Methods, 39(4), 979-984.

Kelley, T. (1935) An unbiased correlation ratio measure. Proceedings of the National Academy of Sciences. 21(9). 554-559.

The computation of CIs is based on the implementation done by Stanley (2018) in the ApaTables package and Kelley (2007) in the MBESS package. All credits go to them.

See also

Examples

# \donttest{
library(effectsize)
mtcars$am_f <- factor(mtcars$am)
mtcars$cyl_f <- factor(mtcars$cyl)

model <- aov(mpg ~ am_f * cyl_f, data = mtcars)

eta_squared(model)

#> Parameter  | Eta_Sq_partial |       90% CI
#> ------------------------------------------
#> am_f       |           0.63 | [0.42, 0.75]
#> cyl_f      |           0.66 | [0.45, 0.77]
#> am_f:cyl_f |           0.10 | [0.00, 0.27]

eta_squared(model, partial = FALSE)

#> Parameter  | Eta_Sq |       90% CI
#> ----------------------------------
#> am_f       |   0.36 | [0.14, 0.54]
#> cyl_f      |   0.41 | [0.16, 0.57]
#> am_f:cyl_f |   0.02 | [0.00, 0.12]

omega_squared(model)

#> Parameter  | Omega_Sq_partial |        90% CI
#> ---------------------------------------------
#> am_f       |             0.57 | [ 0.39, 0.73]
#> cyl_f      |             0.60 | [ 0.40, 0.74]
#> am_f:cyl_f |             0.02 | [-0.07, 0.21]

epsilon_squared(model)

#> Parameter  | Epsilon_Sq_partial |        90% CI
#> -----------------------------------------------
#> am_f       |               0.61 | [ 0.40, 0.74]
#> cyl_f      |               0.63 | [ 0.41, 0.75]
#> am_f:cyl_f |               0.03 | [-0.08, 0.22]

cohens_f(model)

#> Parameter  | Cohens_f_partial |       90% CI
#> --------------------------------------------
#> am_f       |             1.30 | [0.86, 1.73]
#> cyl_f      |             1.38 | [0.90, 1.81]
#> am_f:cyl_f |             0.33 | [0.00, 0.61]

model <- aov(mpg ~ cyl_f * am_f + Error(vs / am_f), data = mtcars)
epsilon_squared(model)

#> Group  |  Parameter | Epsilon_Sq_partial |        90% CI
#> --------------------------------------------------------
#> Within |       am_f |              -0.04 | [-0.04, 0.03]
#> Within | cyl_f:am_f |              -0.03 | [-0.04, 0.10]
#> Within |      cyl_f |               0.40 | [ 0.12, 0.58]

# Recommended:
# Type-3 effect sizes + effects coding
if (require(car, quietly = TRUE)) {
  contrasts(mtcars$am_f) <- contr.sum
  contrasts(mtcars$cyl_f) <- contr.sum

  model <- aov(mpg ~ am_f * cyl_f, data = mtcars)
  model_anova <- car::Anova(model, type = 3)

  eta_squared(model_anova)
}

#> Registered S3 methods overwritten by 'car':
#>   method                          from
#>   influence.merMod                lme4
#>   cooks.distance.influence.merMod lme4
#>   dfbeta.influence.merMod         lme4
#>   dfbetas.influence.merMod        lme4

#> Parameter  | Eta_Sq_partial |       90% CI
#> ------------------------------------------
#> am_f       |           0.11 | [0.00, 0.32]
#> cyl_f      |           0.63 | [0.42, 0.75]
#> am_f:cyl_f |           0.10 | [0.00, 0.27]

if (require("parameters")) {
  data(mtcars)
  model <- lm(mpg ~ wt + cyl, data = mtcars)
  mp <- model_parameters(model)
  eta_squared(mp)
}

#> Loading required package: parameters

#> Parameter   | Eta_Sq_partial |       90% CI
#> -------------------------------------------
#> (Intercept) |           0.95 | [0.92, 0.97]
#> wt          |           0.38 | [0.16, 0.56]
#> cyl         |           0.31 | [0.10, 0.50]

if (require(lmerTest, quietly = TRUE)) {
  model <- lmer(mpg ~ am_f * cyl_f + (1|vs), data = mtcars)
  omega_squared(model)
}

#> 
#> Attaching package: 'lmerTest'

#> The following object is masked from 'package:lme4':
#> 
#>     lmer

#> The following object is masked from 'package:stats':
#> 
#>     step

#> boundary (singular) fit: see ?isSingular

#> Parameter  | Omega_Sq_partial |        90% CI
#> ---------------------------------------------
#> am_f       |             0.07 | [-0.04, 0.28]
#> cyl_f      |             0.60 | [ 0.36, 0.72]
#> am_f:cyl_f |             0.03 | [-0.07, 0.21]

# }