Simulate Eta Squared from Posterior Predictive Distribution

This function simulates data from the posterior predictive distribution (ppd) and for each simulation the Eta Squared is computed for the model's fixed effects.

Effect sizes are computed using the sums of squares obtained from car::Anova(model, ...) which might not always be appropriate (Yeah... ANOVAs are hard...), e.g., when factors aren't 0-mean codded and covariates aren't centered; a warning will be given in such predictors are detected. See details section in eta_squared().

eta_squared_posterior(
  model,
  partial = TRUE,
  type = 3,
  draws = 500,
  verbose = TRUE,
  ...
)

Arguments

model
A Bayesian linear model, fit with brms or rstanarm.

partial
Whether Partial Eta should be returned.

type
Type of sum-of-squares to use. See car::Anova().

draws
An integer indicating the number of draws from the posterior predictive distribution to return. Larger numbers take longer to run, but provide estimates that are more stable.

verbose
Show messages.

...
Currently not used.

Value

A data frame containing the ppd of the Eta squared for each fixed effect, which can then be passed to bayestestR::describe_posterior() for summary stats.

Examples

# \donttest{
if (require(rstanarm) && require(bayestestR)) {
  fit_bayes <- stan_glm(mpg ~ factor(cyl) * wt + qsec,
                        data = mtcars,
                        family = gaussian(),
                        refresh = 0)

  es <- eta_squared_posterior(fit_bayes)

  bayestestR::describe_posterior(es)
}

#> Loading required package: rstanarm

#> Loading required package: Rcpp

#> This is rstanarm version 2.21.1

#> - See https://mc-stan.org/rstanarm/articles/priors for changes to default priors!

#> - Default priors may change, so it's safest to specify priors, even if equivalent to the defaults.

#> - For execution on a local, multicore CPU with excess RAM we recommend calling

#>   options(mc.cores = parallel::detectCores())

#> 
#> Attaching package: ‘rstanarm’

#> The following object is masked from ‘package:car’:
#> 
#>     logit

#> Loading required package: bayestestR

#> Warning: Not all variables are centered:
#>  wt, qsec, cyl
#>  Results might be bogus...

#> Simulating effect size... This can take a while...

#> # Description of Posterior Distributions
#> 
#> Parameter      | Median |         89% CI | pd |        89% ROPE | % in ROPE
#> ---------------------------------------------------------------------------
#> factor(cyl)    |  0.302 | [0.036, 0.551] |  1 | [-0.100, 0.100] |    10.538
#> wt             |  0.459 | [0.192, 0.721] |  1 | [-0.100, 0.100] |     0.000
#> qsec           |  0.140 | [0.000, 0.399] |  1 | [-0.100, 0.100] |    42.601
#> factor(cyl):wt |  0.249 | [0.004, 0.497] |  1 | [-0.100, 0.100] |    16.816
#> 

# compare to:
if (require(car)) {
  fit_freq <- lm(mpg ~ factor(cyl) * wt + qsec,
                 data = mtcars)
  aov_table <- car::Anova(fit_freq, type = 3)
  eta_squared(aov_table)
}

#> Parameter      | Eta2 (partial) |       90% CI
#> ----------------------------------------------
#> factor(cyl)    |           0.29 | [0.05, 0.49]
#> wt             |           0.49 | [0.25, 0.65]
#> qsec           |           0.16 | [0.00, 0.37]
#> factor(cyl):wt |           0.24 | [0.01, 0.43]

# }