Normalization

Performs a normalization of data, i.e., it scales all numeric variables in the range 0 - 1. This is a special case of change_scale().

normalize(x, ...)

# S3 method for numeric
normalize(x, include_bounds = TRUE, verbose = TRUE, ...)

# S3 method for grouped_df
normalize(
  x,
  select = NULL,
  exclude = NULL,
  include_bounds = TRUE,
  verbose = TRUE,
  ...
)

# S3 method for data.frame
normalize(
  x,
  select = NULL,
  exclude = NULL,
  include_bounds = TRUE,
  verbose = TRUE,
  ...
)

Arguments

x
Object.

...
Arguments passed to or from other methods.

include_bounds
Logical, if TRUE, return value may include 0 and 1. If FALSE, the return value is compressed, using the formula (x * (n - 1) + 0.5) / n (Smithson and Verkuilen 2006), to avoid zeros and ones in the normalized variables. This can be useful in case of beta-regression, where the response variable is not allowed to include zeros and ones.

verbose
Toggle warnings on or off.

select
Character vector of column names. If NULL (the default), all variables will be selected.

exclude
Character vector of column names to be excluded from selection.

Value

A normalized object.

References

Smithson M, Verkuilen J (2006). A Better Lemon Squeezer? Maximum-Likelihood Regression with Beta-Distributed Dependent Variables. Psychological Methods, 11(1), 54–71.

See also

Examples

normalize(c(0, 1, 5, -5, -2))

#> [1] 0.5 0.6 1.0 0.0 0.3

normalize(c(0, 1, 5, -5, -2), include_bounds = FALSE)

#> [1] 0.50 0.58 0.90 0.10 0.34

head(normalize(iris))

#>   Sepal.Length Sepal.Width Petal.Length Petal.Width Species
#> 1   0.22222222   0.6250000   0.06779661  0.04166667  setosa
#> 2   0.16666667   0.4166667   0.06779661  0.04166667  setosa
#> 3   0.11111111   0.5000000   0.05084746  0.04166667  setosa
#> 4   0.08333333   0.4583333   0.08474576  0.04166667  setosa
#> 5   0.19444444   0.6666667   0.06779661  0.04166667  setosa
#> 6   0.30555556   0.7916667   0.11864407  0.12500000  setosa