Convert between Odds and Probabilities

Convert between Odds and Probabilities

odds_to_probs(odds, log = FALSE, ...)

# S3 method for data.frame
odds_to_probs(odds, log = FALSE, select = NULL, exclude = NULL, ...)

probs_to_odds(probs, log = FALSE, ...)

# S3 method for data.frame
probs_to_odds(probs, log = FALSE, select = NULL, exclude = NULL, ...)

convert_odds_to_probs(odds, log = FALSE, ...)

convert_probs_to_odds(probs, log = FALSE, ...)

Arguments

odds
The Odds (or log(odds) when log = TRUE) to convert.

log
Take in or output log odds (such as in logistic models).

...
Arguments passed to or from other methods.

select
When a data frame is passed, character or list of of column names to be transformed.

exclude
When a data frame is passed, character or list of column names to be excluded from transformation.

probs
Probability values to convert.

See also

Examples

odds_to_probs(3)

#> [1] 0.75

odds_to_probs(1.09, log = TRUE)

#> [1] 0.7483817

probs_to_odds(0.95)

#> [1] 19

probs_to_odds(0.95, log = TRUE)

#> [1] 2.944439