Convert test statistics (t, z, F) to effect sizes of differences (Cohen's d) or association (partial r)

These functions are convenience functions to convert t, z and F test statistics to Cohen's d and partial r. These are useful in cases where the data required to compute these are not easily available or their computation is not straightforward (e.g., in liner mixed models, contrasts, etc.).
See Effect Size from Test Statistics vignette.

t_to_d(t, df_error, paired = FALSE, ci = 0.95, pooled, ...)

convert_t_to_d(t, df_error, paired = FALSE, ci = 0.95, pooled, ...)

z_to_d(z, n, paired = FALSE, ci = 0.95, pooled, ...)

convert_z_to_d(z, n, paired = FALSE, ci = 0.95, pooled, ...)

F_to_d(f, df, df_error, paired = FALSE, ci = 0.95, pooled, ...)

convert_F_to_d(f, df, df_error, paired = FALSE, ci = 0.95, pooled, ...)

t_to_r(t, df_error, ci = 0.95, ...)

convert_t_to_r(t, df_error, ci = 0.95, ...)

z_to_r(z, n, ci = 0.95, ...)

convert_z_to_r(z, n, ci = 0.95, ...)

F_to_r(f, df, df_error, ci = 0.95, ...)

convert_F_to_r(f, df, df_error, ci = 0.95, ...)

Arguments

t, f, z
The t, the F or the z statistics.

paired
Should the estimate accout for the t-value being testing the difference between dependant means?

ci
Confidence Interval (CI) level

pooled
Deprecated. Use paired.

...
Arguments passed to or from other methods.

n
The number of observations (the sample size).

df, df_error
Degrees of freedom of numerator or of the error estimate (i.e., the residuals).

r
The correlation coefficient r.

Value

A data frame with the effect size(s) between 0-1, and confidence interval(s)

Details

These functions use the following formulae:

$$r_{partial} = t / \sqrt{t^2 + df_{error}}$$

$$r_{partial} = z / \sqrt{z^2 + N}$$

$$Cohen's d = 2 * t / \sqrt{df_{error}}$$

$$Cohen's d_z = t / \sqrt{df_{error}}$$

$$Cohen's d = 2 * z / \sqrt{N}$$

Confidence Intervals

Confidence intervals are estimated using the Noncentrality parameter method; These methods searches for a the best ncp (non-central parameters) for of the noncentral F distribution for the desired tail-probabilities, and then convert these ncps to the corresponding effect sizes.

References

Friedman, H. (1982). Simplified determinations of statistical power, magnitude of effect and research sample sizes. Educational and Psychological Measurement, 42(2), 521-526. doi: 10.1177/001316448204200214

Wolf, F. M. (1986). Meta-analysis: Quantitative methods for research synthesis (Vol. 59). Sage.

Rosenthal, R. (1991). Meta-analytic procedures for social research. Newbury Park, CA: SAGE Publications, Incorporated.

Steiger, J. H. (2004). Beyond the F test: Effect size confidence intervals and tests of close fit in the analysis of variance and contrast analysis. Psychological Methods, 9, 164-182.

Cumming, G., & Finch, S. (2001). A primer on the understanding, use, and calculation of confidence intervals that are based on central and noncentral distributions. Educational and Psychological Measurement, 61(4), 532-574.

Examples

## t Tests
res <- t.test(1:10, y = c(7:20), var.equal = TRUE)
t_to_d(t = res$statistic, res$parameter)

#>     d |         95% CI
#> ----------------------
#> -2.19 | [-3.23, -1.12]

t_to_r(t = res$statistic, res$parameter)

#>     r |         95% CI
#> ----------------------
#> -0.74 | [-0.85, -0.49]

res <- with(sleep, t.test(extra[group == 1], extra[group == 2], paired = TRUE))
t_to_d(t = res$statistic, res$parameter, paired = TRUE)

#>     d |         95% CI
#> ----------------------
#> -1.35 | [-2.23, -0.44]

t_to_r(t = res$statistic, res$parameter)

#>     r |         95% CI
#> ----------------------
#> -0.80 | [-0.91, -0.40]

# \donttest{
## Linear Regression
model <- lm(Sepal.Length ~ Sepal.Width + Petal.Length, data = iris)
library(parameters)
(param_tab <- parameters(model))

#> Parameter    | Coefficient |   SE |       95% CI |     t |  df |      p
#> -----------------------------------------------------------------------
#> (Intercept)  |        2.25 | 0.25 | [1.76, 2.74] |  9.07 | 147 | < .001
#> Sepal.Width  |        0.60 | 0.07 | [0.46, 0.73] |  8.59 | 147 | < .001
#> Petal.Length |        0.47 | 0.02 | [0.44, 0.51] | 27.57 | 147 | < .001

# > Parameter    | Coefficient |   SE |       95% CI |     t |  df |      p
# > -----------------------------------------------------------------------
# > (Intercept)  |        2.25 | 0.25 | [1.76, 2.74] |  9.07 | 147 | < .001
# > Sepal.Width  |        0.60 | 0.07 | [0.46, 0.73] |  8.59 | 147 | < .001
# > Petal.Length |        0.47 | 0.02 | [0.44, 0.51] | 27.57 | 147 | < .001

t_to_r(param_tab$t[2:3], param_tab$df_error[2:3])

#>    r |       95% CI
#> -------------------
#> 0.58 | [0.47, 0.66]
#> 0.92 | [0.89, 0.93]

# > [1] 0.5781005 0.9153894

## Use with emmeans based contrasts (see also t_to_eta2)
if (require(emmeans) & require(dplyr)) {
  warp.lm <- lm(breaks ~ wool * tension, data = warpbreaks)

  emmeans(warp.lm,  ~ tension | wool) %>%
    contrast("pairwise") %>%
    summary() %>%
    bind_cols(t_to_d(.$t.ratio, .$df))
}

#> wool = A:
#>  contrast estimate   SE df t.ratio p.value       d   CI  CI_low CI_high
#>  L - M      20.556 5.16 48  3.986  0.0007   1.1506 0.95  0.5348   1.756
#>  L - H      20.000 5.16 48  3.878  0.0009   1.1195 0.95  0.5061   1.723
#>  M - H      -0.556 5.16 48 -0.108  0.9936  -0.0311 0.95 -0.5968   0.535
#> 
#> wool = B:
#>  contrast estimate   SE df t.ratio p.value       d   CI  CI_low CI_high
#>  L - M      -0.556 5.16 48 -0.108  0.9936  -0.0311 0.95 -0.5968   0.535
#>  L - H       9.444 5.16 48  1.831  0.1704   0.5286 0.95 -0.0496   1.102
#>  M - H      10.000 5.16 48  1.939  0.1389   0.5597 0.95 -0.0198   1.134
#> 
#> P value adjustment: tukey method for comparing a family of 3 estimates 

# How does this compare to actual partial correlations?
if (require("ppcor")) {
  pcor(iris[1:3])$estimate[1, -1]
}

#> Loading required package: ppcor

#> Loading required package: MASS

#> 
#> Attaching package: 'MASS'

#> The following object is masked from 'package:dplyr':
#> 
#>     select

#>  Sepal.Width Petal.Length 
#>    0.5781005    0.9153894 

# }