Adjust data for the effect of other variable(s)

This function can be used to adjust the data for the effect of other variables present in the dataset. It is based on an underlying fitting of regressions models, allowing for quite some flexibility, such as including factors as random effects in mixed models (multilevel partialization), continuous variables as smooth terms in general additive models (non-linear partialization) and/or fitting these models under a Bayesian framework. The values returned by this function are the residuals of the regression models. Note that a regular correlation between two "adjusted" variables is equivalent to the partial correlation between them.

adjust(
  data,
  effect = NULL,
  select = NULL,
  exclude = NULL,
  multilevel = FALSE,
  additive = FALSE,
  bayesian = FALSE
)

Arguments

data
A dataframe.

effect
Character vector of column names to be adjusted for (regressed out). If NULL (the default), all variables will be selected.

select
Character vector of column names. If NULL (the default), all variables will be selected.

exclude
Character vector of column names to be excluded from selection.

multilevel
If TRUE, the factors are included as random factors. Else, if FALSE (default), they are included as fixed effects in the simple regression model.

additive
If TRUE, continuous variables as included as smooth terms in additive models. The goal is to regress-out potential non-linear effects.

bayesian
If TRUE, the models are fitted under the Bayesian framework using rstanarm.

Value

A data frame comparable to data, with adjusted variables.

Examples

adjust(attitude)

#>         rating  complaints  privileges     learning      raises    critical
#> 1   -8.1102953   5.5583770 -15.8489495  -2.75102306   0.5742664  15.6055022
#> 2    1.6472337   0.0646564  -1.4225924  -3.06207012  -1.5567655  -2.3157810
#> 3    1.0605589  -7.5116953  11.1746085   5.59808033   4.8603132   8.0618012
#> 4   -0.2268416   3.8345277  -4.5674406   0.03866933  -7.1185324  13.0025744
#> 5    6.5462010  -1.2420122  -3.0510984   0.87312095  -2.7131349   6.5003528
#> 6  -10.9418499   5.2030745   2.6641564  -1.24552098   4.1370346 -21.6783822
#> 7   -9.1484140   5.4703108 -10.4328432   5.71859149   5.3656703  -5.2384340
#> 8    0.9029929   2.5608484  -5.0517951  -5.29374027   4.4574753 -11.0030328
#> 9   -7.5309862   2.4339896  12.2619091  10.26440961   2.8523548   9.9475629
#> 10   7.8015424  -4.3358833  -2.6727182  -7.23210392  -0.3002972   4.0961127
#> 11   6.0742817 -11.9900391   8.0975792   5.47869539   2.3150922  -3.6824794
#> 12  11.5989723  -4.8671195   1.5027728 -15.87942464  -0.7291924  -2.7471720
#> 13   9.4183197  -7.3970919  13.2328606  -9.42000500   2.1992437  -8.9264218
#> 14  -2.2140147   8.1881193  24.9814055 -11.64285330  -6.1495840   2.8418688
#> 15   0.4506705  -2.5770669  -5.2547496   6.64133914   6.4784591  -0.8454553
#> 16  -3.5478519  14.2969120 -14.9748695   9.70538089  -9.5300033 -13.8204493
#> 17  -2.1501319   7.2584531  -2.9433798  -2.89172149  -1.5873728  -2.1520580
#> 18   3.6026355  -3.3274750   6.7012942  12.70306315 -16.4991421  10.2712762
#> 19  -3.0165587   1.5321885  -8.8574173  -0.23273381   5.8698681   5.8895747
#> 20  -5.6201442  -1.3898277  15.5623553  -1.18623905   0.3805922   1.6696336
#> 21   7.3967582  -5.9625957  -3.2550843  -7.29972652  -4.0965067  -5.2607986
#> 22   0.1809831  -5.5453037   0.9259830   7.35994049   2.2144741   6.2047767
#> 23 -10.6639999   6.9119296  -0.5164097   2.31172066   1.3676495   6.2435385
#> 24  -4.6247464  -5.3189073  -1.9093845  10.68079101  -2.6323079 -11.5620146
#> 25   5.6828983 -10.8499327  -0.5640085  -2.57618485   9.8253009  -0.8953311
#> 26  -1.8434727   1.4351539   2.6267589 -11.80931691   8.8582382 -11.7425336
#> 27   2.8596385  -6.1038672  -0.9715536   6.42826587   6.7173702  -0.5581726
#> 28  -8.0453540   7.8050105  -4.9863879   2.81595072  -7.8541569  12.9751638
#> 29   7.3394730  -1.7770932   7.0405122  -3.58130062  -0.5856674  -5.4892951
#> 30   5.1215016   7.6423593 -19.4915135  -0.51405450  -7.1207394   4.6080728
#>        advance
#> 1    2.8684130
#> 2    5.3937097
#> 3   -6.4236221
#> 4   -0.3951046
#> 5    2.1988621
#> 6   -3.1912418
#> 7   -8.2975047
#> 8    0.3313834
#> 9  -17.8173479
#> 10   3.2077674
#> 11  -7.7480571
#> 12   8.2201619
#> 13  -5.8982920
#> 14  -0.3236635
#> 15  -5.7639676
#> 16   0.4424158
#> 17  11.6172195
#> 18  11.7550726
#> 19  -2.2362899
#> 20   2.7752634
#> 21   5.0639714
#> 22  -6.4649197
#> 23  -6.3462052
#> 24   4.3270326
#> 25  -8.6992680
#> 26  14.8238571
#> 27  -4.6986777
#> 28   0.5363293
#> 29   6.7021863
#> 30   4.0405164

adjust(attitude, effect = "complaints", select = "rating")

#>          rating complaints privileges learning raises critical advance
#> 1   -9.86142016         51         30       39     61       92      45
#> 2    0.32865220         64         51       54     63       73      47
#> 3    3.80099328         70         68       69     76       86      48
#> 4   -0.91673799         63         45       47     54       84      35
#> 5    7.76411473         78         56       66     71       83      47
#> 6  -12.87985944         55         49       44     54       49      34
#> 7   -6.93517726         67         42       56     66       68      35
#> 8    0.02794419         75         50       55     70       66      41
#> 9   -4.25432454         82         72       67     71       83      31
#> 10   6.59248165         61         45       47     62       80      41
#> 11   9.62936020         53         53       58     58       67      34
#> 12   7.34709147         60         47       39     59       74      41
#> 13   7.83787183         62         57       42     55       63      25
#> 14  -9.00893436         83         83       45     59       77      35
#> 15   4.51872455         77         54       72     79       77      46
#> 16  -1.29120309         90         50       72     60       54      36
#> 17  -4.51815400         85         64       69     79       79      63
#> 18   5.34709147         60         65       75     55       80      60
#> 19  -2.19900672         70         46       57     75       85      46
#> 20  -8.14368889         58         68       54     64       78      52
#> 21   5.43928784         40         33       34     43       64      33
#> 22   3.59248165         61         52       62     66       80      41
#> 23 -11.18056744         66         52       50     63       80      37
#> 24  -2.29688270         37         42       58     50       57      49
#> 25   7.87475038         54         42       48     66       75      33
#> 26  -6.48127545         77         66       63     88       76      72
#> 27   7.02794419         75         58       74     80       78      49
#> 28  -9.38907907         57         44       45     51       83      38
#> 29   6.48184600         85         71       71     77       74      55
#> 30   5.74567546         82         39       59     64       78      39

# \donttest{
if (FALSE) {
adjust(attitude, effect = "complaints", select = "rating", bayesian = TRUE)
adjust(attitude, effect = "complaints", select = "rating", additive = TRUE)
attitude$complaints_LMH <- cut(attitude$complaints, 3)
adjust(attitude, effect = "complaints_LMH", select = "rating", multilevel = TRUE)
}
# }