Conversion between Effect sizes for Contingency Tables (Chi2, Phi, Cramer's V...)

Convert between Chi square, (\(\chi^2\)), phi (\(\phi\)) and Cramer's V.

chisq_to_phi(chisq, n, nrow, ncol, ci = 0.95, adjust = FALSE, ...)

chisq_to_cramers_v(chisq, n, nrow, ncol, ci = 0.95, adjust = FALSE, ...)

phi_to_chisq(phi, n, ...)

Arguments

chisq
The Chi2 statistic.

n
Sample size.

nrow, ncol
The number of rows/columns in the contingency table (ignored for Phi when adjust=FALSE and CI=NULL).

ci
Confidence Interval (CI) level

adjust
Should the effect size be bias-corrected? Defaults to FALSE.

...
Arguments passed to or from other methods.

phi
The Phi statistic.

Value

A data frame with the effect size(s) between 0-1, and confidence interval(s).

Details

These functions use the following formulae:
$$\phi = \sqrt{\chi^2 / n}$$
$$Cramer's V = \phi / \sqrt{min(nrow,ncol)-1}$$
For adjusted versions, see Bergsma, 2013.

Confidence Intervals

Confidence intervals are estimated using the Noncentrality parameter method; These methods searches for a the best ncp (non-central parameters) for of the noncentral F distribution for the desired tail-probabilities, and then convert these ncps to the corresponding effect sizes.

References

Cumming, G., & Finch, S. (2001). A primer on the understanding, use, and calculation of confidence intervals that are based on central and noncentral distributions. Educational and Psychological Measurement, 61(4), 532-574.

Bergsma, W. (2013). A bias-correction for Cramer's V and Tschuprow's T. Journal of the Korean Statistical Society, 42(3), 323-328.

Examples

contingency_table <- as.table(rbind(c(762, 327, 468), c(484, 239, 477), c(484, 239, 477)))

chisq.test(contingency_table)

#> 
#> 	Pearson's Chi-squared test
#> 
#> data:  contingency_table
#> X-squared = 41.234, df = 4, p-value = 2.405e-08
#> 

#
#         Pearson's Chi-squared test
#
# data:  ctab
# X-squared = 41.234, df = 4, p-value = 2.405e-08

chisq_to_phi(41.234,
  n = sum(contingency_table),
  nrow = nrow(contingency_table),
  ncol = ncol(contingency_table)
)

#>  Phi |       95% CI
#> -------------------
#> 0.10 | [0.07, 0.13]

chisq_to_cramers_v(41.234,
  n = sum(contingency_table),
  nrow = nrow(contingency_table),
  ncol = ncol(contingency_table)
)

#> Cramer's V |       95% CI
#> -------------------------
#>       0.07 | [0.05, 0.09]