Effect size for differences

Compute effect size indices for standardized differences: Cohen's d, Hedges' g and Glass’s delta. (This function returns the population estimate.)

Both Cohen's d and Hedges' g are the estimated the standardized difference between the means of two populations. Hedges' g provides a bias correction to Cohen's d for small sample sizes. For sample sizes > 20, the results for both statistics are roughly equivalent. Glass’s delta is appropriate when the standard deviations are significantly different between the populations, as it uses only the second group's standard deviation.

cohens_d(
  x,
  y = NULL,
  data = NULL,
  pooled_sd = TRUE,
  paired = FALSE,
  ci = 0.95,
  correction
)

hedges_g(
  x,
  y = NULL,
  data = NULL,
  correction = 1,
  pooled_sd = TRUE,
  paired = FALSE,
  ci = 0.95
)

glass_delta(x, y = NULL, data = NULL, ci = 0.95, correction)

# S3 method for effectsize_difference
print(x, digits = 2, append_CL = FALSE, ...)

Arguments

x
A formula, a numeric vector, or a character name of one in data. (For print() the result of one of the standardized difference functions.)

y
A numeric vector, a grouping (character / factor) vector, a or a character name of one in data. Ignored if x is a formula.

data
An optional data frame containing the variables.

pooled_sd
If TRUE (default), a sd_pooled() is used (assuming equal variance). Else the mean SD from both groups is used instead.

paired
If TRUE, the values of x and y are considered as paired. This produces an effect size that is equivalent to the one-sample effect size on x - y.

ci
Confidence Interval (CI) level

correction
Type of small sample bias correction to apply to produce Hedges' g. Can be 1 for Hedges and Olkin's original correction (default) or 2 for Hunter and Schmidt's correction (see McGrath & Meyer, 2006).

digits
Number of significant digits.

append_CL
Should the Common Language Effect Sizes be printed as well? Not applicable to Glass' Delta (See d_to_common_language())

...
Not used.

Value

A data frame with the effect size ( Cohens_d, Hedges_g, Glass_delta) and their CIs (CI_low and CI_high).

Details

Note

The indices here give the population estimated standardized difference. Some statistical packages give the sample estimate instead (without applying Bessel's correction).

Confidence Intervals

Confidence intervals are estimated using the Noncentrality parameter method; These methods searches for a the best non-central parameters (ncps) of the noncentral t-, F- or Chi-squared distribution for the desired tail-probabilities, and then convert these ncps to the corresponding effect sizes. (See full effectsize-CIs for more.)

References

Cohen, J. (1988). Statistical power analysis for the behavioral sciences (2nd Ed.). New York: Routledge.

Hedges, L. V. & Olkin, I. (1985). Statistical methods for meta-analysis. Orlando, FL: Academic Press.

Hunter, J. E., & Schmidt, F. L. (2004). Methods of meta-analysis: Correcting error and bias in research findings. Sage.

McGrath, R. E., & Meyer, G. J. (2006). When effect sizes disagree: the case of r and d. Psychological methods, 11(4), 386.

See also

Examples

cohens_d(sleep$extra, sleep$group)

#> Cohen's d |        95% CI
#> -------------------------
#>     -0.83 | [-1.74, 0.10]
#> 
#> - Estimated using pooled SD.

hedges_g("extra", "group", data = sleep)

#> Hedge's g |        95% CI
#> -------------------------
#>     -0.80 | [-1.67, 0.09]
#> 
#> - Estimated using pooled SD.
#> - Bias corrected using Hedges and Olkin's method.

cohens_d(sleep$extra[sleep$group==1], sleep$extra[sleep$group==2], paired = TRUE)

#> Cohen's d |         95% CI
#> --------------------------
#>     -1.28 | [-2.23, -0.44]
#> 
#> - Estimated using pooled SD.

cohens_d(mpg ~ am, data = mtcars)

#> Cohen's d |         95% CI
#> --------------------------
#>     -1.48 | [-2.27, -0.67]
#> 
#> - Estimated using pooled SD.

cohens_d(mpg ~ am, data = mtcars, pooled_sd = FALSE)

#> Cohen's d |         95% CI
#> --------------------------
#>     -1.41 | [-2.17, -0.51]
#> 
#> - Estimated using un-pooled SD.

hedges_g(mpg ~ am, data = mtcars)

#> Hedge's g |         95% CI
#> --------------------------
#>     -1.44 | [-2.21, -0.65]
#> 
#> - Estimated using pooled SD.
#> - Bias corrected using Hedges and Olkin's method.

glass_delta(mpg ~ am, data = mtcars)

#> Glass' delta |         95% CI
#> -----------------------------
#>        -1.17 | [-1.95, -0.46]
#> 
#> - Estimated using pooled SD.

print(cohens_d(mpg ~ am, data = mtcars), append_CL = TRUE)

#> Cohen's d |         95% CI
#> --------------------------
#>     -1.48 | [-2.27, -0.67]
#> 
#> - Estimated using pooled SD.
#> 
#> # Common Language Effect Sizes
#> 
#> Cohen's U3 | Overlap | Probability of superiority
#> -------------------------------------------------
#> 6.97%      |  45.99% |                     14.80%