Convert between d, r and Odds ratio

Enables a conversion between different indices of effect size, such as standardized difference (Cohen's d), correlation r or (log) odds ratios.

d_to_r(d, ...)

r_to_d(r, ...)

convert_d_to_r(d, ...)

convert_r_to_d(r, ...)

oddsratio_to_d(OR, log = FALSE, ...)

convert_oddsratio_to_d(OR, log = FALSE, ...)

logoddsratio_to_d(OR, log = TRUE, ...)

d_to_oddsratio(d, log = FALSE, ...)

convert_d_to_oddsratio(d, log = FALSE, ...)

oddsratio_to_r(OR, log = FALSE, ...)

convert_oddsratio_to_r(OR, log = FALSE, ...)

logoddsratio_to_r(OR, log = TRUE, ...)

r_to_oddsratio(r, log = FALSE, ...)

convert_r_to_oddsratio(r, log = FALSE, ...)

Arguments

d
Standardized difference value (Cohen's d).

...
Arguments passed to or from other methods.

r
Correlation coefficient r.

OR
Odds ratio values in vector or data frame.

log
Take in or output the log of the ratio (such as in logistic models).

Value

Converted index.

Details

Conversions between OR and r is done through these formulae.

d to r: \(d = \frac{2 * r}{\sqrt{1 - r^2}}\)

r to d: \(r = \frac{d}{\sqrt{d^2 + 4}}\)

OR to d: \(d = \frac{\log(OR)\times\sqrt{3}}{\pi}\)

d to OR: \(log(OR) = d * \frac{\pi}{\sqrt(3)}\)

The conversion from d to r assumes equally sized groups. The resulting r is also called the binomial effect size display (BESD; Rosenthal et al., 1982).

References

Sánchez-Meca, J., Marín-Martínez, F., & Chacón-Moscoso, S. (2003). Effect-size indices for dichotomized outcomes in meta-analysis. Psychological methods, 8(4), 448.

Borenstein, M., Hedges, L. V., Higgins, J. P. T., & Rothstein, H. R. (2009). Converting among effect sizes. Introduction to meta-analysis, 45-49.

Rosenthal, R., & Rubin, D. B. (1982). A simple, general purpose display of magnitude of experimental effect. Journal of educational psychology, 74(2), 166.

See also

Examples

r_to_d(0.5)

#> [1] 1.154701

d_to_oddsratio(1.154701)

#> [1] 8.120534

oddsratio_to_r(8.120534)

#> [1] 0.5000001

d_to_r(1)

#> [1] 0.4472136

r_to_oddsratio(0.4472136, log = TRUE)

#> [1] 1.813799

oddsratio_to_d(1.813799, log = TRUE)

#> [1] 0.9999998