General effect size conversion

Enables a conversion between different indices of effect size, such as standardized difference (Cohen's d), correlation r or (log) odds ratios.

percentage_to_d(percentage, ...)

d_to_percentage(d, ...)

convert_percentage_to_d(percentage, ...)

convert_d_to_percentage(d, ...)

d_to_r(d, ...)

r_to_d(r, ...)

convert_d_to_r(d, ...)

convert_r_to_d(r, ...)

odds_to_d(odds, log = FALSE, ...)

convert_odds_to_d(odds, log = FALSE, ...)

logodds_to_d(odds, log = TRUE, ...)

d_to_odds(d, log = FALSE, ...)

convert_d_to_odds(d, log = FALSE, ...)

odds_to_r(odds, log = FALSE, ...)

convert_odds_to_r(odds, log = FALSE, ...)

logodds_to_r(odds, log = TRUE, ...)

r_to_odds(r, log = FALSE, ...)

convert_r_to_odds(r, log = FALSE, ...)

odds_to_probs(odds, log = FALSE, ...)

# S3 method for data.frame
odds_to_probs(odds, log = FALSE, select = NULL, exclude = NULL, ...)

probs_to_odds(probs, log = FALSE, ...)

convert_odds_to_probs(odds, log = FALSE, ...)

convert_probs_to_odds(probs, log = FALSE, ...)

Arguments

percentage
Percentage value (e.g., 0.01 for one percent).

...
Arguments passed to or from other methods.

d
Standardized difference value (Cohen's d).

r
Correlation coefficient r.

odds
Odds values in vector or dataframe.

log
Take in or output log odds (such as in logistic models).

select
When a dataframe is passed, character or list of of column names to be transformed.

exclude
When a dataframe is passed, character or list of column names to be excluded from transformation.

probs
Probability values.

Value

Converted index.

Details

d to r: \(d = \frac{2 * r}{\sqrt{1 - r^2}}\)

r to d: \(r = \frac{d}{\sqrt{d^2 + 4}}\)

odds to d: \(d = \frac{\log(odds)\times\sqrt{3}}{\pi}\)

d to odds: \(log(odds) = d * \frac{\pi}{\sqrt(3)}\)

References

Sánchez-Meca, J., Marín-Martínez, F., & Chacón-Moscoso, S. (2003). Effect-size indices for dichotomized outcomes in meta-analysis. Psychological methods, 8(4), 448.

Borenstein, M., Hedges, L. V., Higgins, J. P. T., & Rothstein, H. R. (2009). Converting among effect sizes. Introduction to meta-analysis, 45-49.

Examples

r_to_d(0.5)

#> [1] 1.154701

d_to_odds(d = 1.154701)

#> [1] 8.120534

odds_to_r(odds = 8.120534)

#> [1] 0.5000001

d_to_r(d = 1)

#> [1] 0.4472136

r_to_odds(0.4472136, log = TRUE)

#> [1] 1.813799

odds_to_d(1.813799, log = TRUE)

#> [1] 0.9999998