Effect Size

This function tries to return the best effect-size measure for the provided input model. See details.

effectsize(model, ...)

Arguments

model
An object of class htest, or a statistical model. See details.

...
Arguments passed to or from other methods.

Details

For an object of class htest:

A t-test returns Cohen's d via t_to_d().

A correlation test returns r. See t_to_r().

A Chi-squared test returns Cramer's V via cramers_v().

A One-way ANOVA test returns Eta squared via F_to_eta2(), but can be changes via an es argument.

Objects of class anova, aov, or aovlist are passed to eta_squared().

Other objects are passed to standardize_parameters().

For statistical models it is recommended to directly use the listed functions, for the full range of options they provide.

Examples

contingency_table <- as.table(rbind(c(762, 327, 468), c(484, 239, 477), c(484, 239, 477)))
Xsq <- chisq.test(contingency_table)
effectsize(Xsq)

#> Cramer's V |       95% CI
#> -------------------------
#>       0.07 | [0.05, 0.09]

Ts <- t.test(1:10, y = c(7:20))
effectsize(Ts)

#>     d |         95% CI
#> ----------------------
#> -2.32 | [-3.38, -1.22]

Aov <- oneway.test(extra ~ group, data = sleep)
effectsize(Aov)

#> Eta2 (partial) |       90% CI
#> -----------------------------
#>           0.16 | [0.00, 0.42]

fit <- lm(mpg ~ factor(cyl) * wt + hp, data = mtcars)
effectsize(fit)

#> Parameter       | Coefficient (std.) |         95% CI
#> -----------------------------------------------------
#> (Intercept)     |           1.14e-03 | [-0.53,  0.53]
#> factor(cyl)6    |              -0.20 | [-0.79,  0.38]
#> factor(cyl)8    |              -0.32 | [-1.06,  0.42]
#> wt              |              -0.90 | [-1.33, -0.46]
#> hp              |              -0.25 | [-0.52,  0.01]
#> factor(cyl)6:wt |               0.37 | [-0.63,  1.36]
#> factor(cyl)8:wt |               0.54 | [ 0.03,  1.06]
#> 
#> # Standardization method: Refit
#> 

anova_table <- anova(fit)
effectsize(anova_table)

#> Parameter      | Eta2 (partial) |       90% CI
#> ----------------------------------------------
#> factor(cyl)    |           0.86 | [0.76, 0.91]
#> wt             |           0.47 | [0.22, 0.64]
#> hp             |           0.14 | [0.00, 0.36]
#> factor(cyl):wt |           0.16 | [0.00, 0.30]