eta2_to_f2 {effectsize}	R Documentation

Convert Between ANOVA Effect Sizes

Description

Convert Between ANOVA Effect Sizes

Usage


eta2_to_f2(es)

eta2_to_f(es)

f2_to_eta2(f2)

f_to_eta2(f)

Arguments

es
Any measure of variance explained such as Eta-, Epsilon-, Omega-, or R-Squared, partial or otherwise. See details.

f, f2
Cohen's f or f-squared.

Details

Any measure of variance explained can be converted to a corresponding Cohen's f via:

f^2 = \frac{η^2}{1 - η^2}

η^2 = \frac{f^2}{1 + f^2}

If a partial Eta-Squared is used, the resulting Cohen's f is a partial-Cohen's f; If a less biased estimate of variance explained is used (such as Epsilon- or Omega-Squared), the resulting Cohen's f is likewise a less biased estimate of Cohen's f.

References

Cohen, J. (1988). Statistical power analysis for the behavioral sciences (2nd Ed.). New York: Routledge.

Steiger, J. H. (2004). Beyond the F test: Effect size confidence intervals and tests of close fit in the analysis of variance and contrast analysis. Psychological Methods, 9, 164-182.

See Also

eta_squared() for more details.

Other convert between effect sizes: d_to_r(), diff_to_cles, odds_to_probs(), oddsratio_to_riskratio(), w_to_fei()

[Package effectsize version 0.8.7 Index]