Parameters standardization

Compute standardized model parameters (coefficients).

standardize_parameters(
  model,
  parameters = NULL,
  method = "refit",
  robust = FALSE,
  two_sd = FALSE,
  verbose = TRUE,
  centrality = "median",
  ...
)

standardize_posteriors(
  model,
  method = "refit",
  robust = FALSE,
  two_sd = FALSE,
  verbose = TRUE,
  ...
)

Arguments

model
A statistical model.

parameters
An optional table containing the parameters to standardize. If NULL, will automatically retrieve it from the model.

method
The method used for standardizing the parameters. Can be "refit" (default), "posthoc", "smart" or "basic". See 'Details'.

robust
Logical, if TRUE, centering is done by subtracting the median from the variables and dividing it by the median absolute deviation (MAD). If FALSE, variables are standardized by subtracting the mean and dividing it by the standard deviation (SD).

two_sd
If TRUE, the variables are scaled by two times the deviation (SD or MAD depending on robust). This method can be useful to obtain model coefficients of continuous parameters comparable to coefficients related to binary predictors (Gelman, 2008).

verbose
Toggle warnings on or off.

centrality
For Bayesian models, which point-estimates (centrality indices) to compute. Character (vector) or list with one or more of these options: "median", "mean", "MAP" or "all".

...
Arguments passed to or from other methods.

Value

Standardized parameters.

Details

Methods:

refit: This method is based on a complete model re-fit with a standardized version of data. Hence, this method is equal to standardizing the variables before fitting the model. It is the "purest" and the most accurate (Neter et al., 1989), but it is also the most computationally costly and long (especially for heavy models such as, for instance, for Bayesian models). This method is particularly recommended for complex models that include interactions or transformations (e.g., polynomial or spline terms). The robust (default to FALSE) argument enables a robust standardization of data, i.e., based on the median and MAD instead of the mean and SD.

posthoc: Post-hoc standardization of the parameters, aiming at emulating the results obtained by "refit" without refitting the model. The coefficients are divided by the standard deviation (or MAD if robust) of the outcome (which becomes their expression 'unit'). Then, the coefficients related to numeric variables are additionally multiplied by the standard deviation (or MAD if robust) of the related terms, so that they correspond to changes of 1 SD of the predictor (e.g., "A change in 1 SD of x is related to a change of 0.24 of the SD of y). This does not apply to binary variables or factors, so the coefficients are still related to changes in levels. This method is not accurate and tend to give aberrant results when interactions are specified.

smart (Standardization of Model's parameters with Adjustment, Reconnaissance and Transformation): Similar to method = "posthoc" in that it does not involve model refitting. The difference is that the SD of the response is computed on the relevant section of the data. For instance, if a factor with 3 levels A (the intercept), B and C is entered as a predictor, the effect corresponding to B vs. A will be scaled by the variance of the response at the intercept only. As a results, the coefficients for effects of factors are similar to a Glass' delta.

basic: This method is similar to method = "posthoc", but treats all variables as continuous: it also scales the coefficient by the standard deviation of model's matrix' parameter of factors levels (transformed to integers) or binary predictors. Although being inappropriate for these cases, this method is the one implemented by default in other software packages, such as lm.beta::lm.beta().

When method = "smart" or method = "classic", standardize_parameters() also returns the standard errors for the standardized coefficients. Then, ci() can be used to calculate confidence intervals for the standardized coefficients. See 'Examples'.

References

Neter, J., Wasserman, W., & Kutner, M. H. (1989). Applied linear regression models.

Gelman, A. (2008). Scaling regression inputs by dividing by two standard deviations. Statistics in medicine, 27(15), 2865-2873.

See also

Examples

library(effectsize)
data(iris)

model <- lm(Sepal.Length ~ Species * Petal.Width, data = iris)
standardize_parameters(model, method = "refit")

#>                       Parameter Std_Coefficient
#> 1                   (Intercept)      0.05969491
#> 2             Speciesversicolor     -0.16597410
#> 3              Speciesvirginica      0.18985849
#> 4                   Petal.Width      0.85622714
#> 5 Speciesversicolor:Petal.Width      0.45674637
#> 6  Speciesvirginica:Petal.Width     -0.25713539

# \donttest{
standardize_parameters(model, method = "posthoc")

#>                       Parameter Std_Coefficient
#> 1                   (Intercept)       0.0000000
#> 2             Speciesversicolor      -0.8846360
#> 3              Speciesvirginica       0.5944450
#> 4                   Petal.Width       0.8562271
#> 5 Speciesversicolor:Petal.Width       0.4567464
#> 6  Speciesvirginica:Petal.Width      -0.2571354

standardize_parameters(model, method = "smart")

#>                       Parameter Std_Coefficient
#> 1                   (Intercept)       0.0000000
#> 2             Speciesversicolor      -2.0781802
#> 3              Speciesvirginica       1.3964656
#> 4                   Petal.Width       0.8562271
#> 5 Speciesversicolor:Petal.Width       1.0670272
#> 6  Speciesvirginica:Petal.Width      -0.6007064

standardize_parameters(model, method = "basic")

#>                       Parameter Std_Coefficient
#> 1                   (Intercept)       0.0000000
#> 2             Speciesversicolor      -0.4184185
#> 3              Speciesvirginica       0.2811628
#> 4                   Petal.Width       0.8562271
#> 5 Speciesversicolor:Petal.Width       0.3819094
#> 6  Speciesvirginica:Petal.Width      -0.3276012

# Robust and 2 SD
standardize_parameters(model, robust = TRUE)

#>                       Parameter Std_Coefficient
#> 1                   (Intercept)      0.17960924
#> 2             Speciesversicolor     -0.08429934
#> 3              Speciesvirginica      0.12439049
#> 4                   Petal.Width      0.93017273
#> 5 Speciesversicolor:Petal.Width      0.49619195
#> 6  Speciesvirginica:Petal.Width     -0.27934215

standardize_parameters(model, two_sd = TRUE)

#>                       Parameter Std_Coefficient
#> 1                   (Intercept)      0.05969491
#> 2             Speciesversicolor     -0.16597410
#> 3              Speciesvirginica      0.18985849
#> 4                   Petal.Width      0.85622714
#> 5 Speciesversicolor:Petal.Width      0.45674637
#> 6  Speciesvirginica:Petal.Width     -0.25713539

# show CI
library(parameters)
params <- standardize_parameters(model, method = "smart", robust = TRUE)
ci(params)

#>                       Parameter CI     CI_low   CI_high
#> 1                   (Intercept) 95  0.0000000 0.0000000
#> 2             Speciesversicolor 95 -5.7427584 0.8018635
#> 3              Speciesvirginica 95 -1.8953532 5.2154647
#> 4                   Petal.Width 95 -0.3421105 2.2024559
#> 5 Speciesversicolor:Petal.Width 95 -3.3091775 6.7825212
#> 6  Speciesvirginica:Petal.Width 95 -5.7472488 3.7918537

iris$binary <- ifelse(iris$Sepal.Width > 3, 1, 0)
model <- glm(binary ~ Species * Sepal.Length, data = iris, family = "binomial")
standardize_parameters(model, method = "refit")

#> Waiting for profiling to be done...

#>                        Parameter Std_Coefficient
#> 1                    (Intercept)        6.786380
#> 2              Speciesversicolor       -9.544176
#> 3               Speciesvirginica       -8.145297
#> 4                   Sepal.Length        4.238719
#> 5 Speciesversicolor:Sepal.Length       -1.518858
#> 6  Speciesvirginica:Sepal.Length       -3.514712

standardize_parameters(model, method = "posthoc")

#>                        Parameter Std_Coefficient
#> 1                    (Intercept)        0.000000
#> 2              Speciesversicolor        1.173802
#> 3               Speciesvirginica       16.656626
#> 4                   Sepal.Length        4.238719
#> 5 Speciesversicolor:Sepal.Length       -1.518858
#> 6  Speciesvirginica:Sepal.Length       -3.514712

standardize_parameters(model, method = "smart")

#>                        Parameter Std_Coefficient
#> 1                    (Intercept)        0.000000
#> 2              Speciesversicolor        1.173802
#> 3               Speciesvirginica       16.656626
#> 4                   Sepal.Length        4.238719
#> 5 Speciesversicolor:Sepal.Length       -1.518858
#> 6  Speciesvirginica:Sepal.Length       -3.514712

standardize_parameters(model, method = "basic")

#>                        Parameter Std_Coefficient
#> 1                    (Intercept)       0.0000000
#> 2              Speciesversicolor       0.5551894
#> 3               Speciesvirginica       7.8783138
#> 4                   Sepal.Length       4.2387192
#> 5 Speciesversicolor:Sepal.Length      -5.1783651
#> 6  Speciesvirginica:Sepal.Length     -13.3161343

# }

# \donttest{
if (require("rstanarm")) {
  model <- stan_glm(Sepal.Length ~ Species * Petal.Width, data = iris, iter = 500, refresh = 0)
  standardize_posteriors(model, method = "refit")
  standardize_posteriors(model, method = "posthoc")
  standardize_posteriors(model, method = "smart")
  standardize_posteriors(model, method = "basic")

  standardize_parameters(model, method = "refit")
  standardize_parameters(model, method = "posthoc")
  standardize_parameters(model, method = "smart")
  standardize_parameters(model, method = "basic")
}

#> Loading required package: rstanarm

#> Loading required package: Rcpp

#> rstanarm (Version 2.19.3, packaged: 2020-02-11 05:16:41 UTC)

#> - Do not expect the default priors to remain the same in future rstanarm versions.

#> Thus, R scripts should specify priors explicitly, even if they are just the defaults.

#> - For execution on a local, multicore CPU with excess RAM we recommend calling

#> options(mc.cores = parallel::detectCores())

#> - bayesplot theme set to bayesplot::theme_default()

#>    * Does _not_ affect other ggplot2 plots

#>    * See ?bayesplot_theme_set for details on theme setting

#> Warning: Bulk Effective Samples Size (ESS) is too low, indicating posterior means and medians may be unreliable.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#bulk-ess

#> Warning: Tail Effective Samples Size (ESS) is too low, indicating posterior variances and tail quantiles may be unreliable.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#tail-ess

#> Warning: The largest R-hat is 1.06, indicating chains have not mixed.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#r-hat

#> Warning: Bulk Effective Samples Size (ESS) is too low, indicating posterior means and medians may be unreliable.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#bulk-ess

#> Warning: Tail Effective Samples Size (ESS) is too low, indicating posterior variances and tail quantiles may be unreliable.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#tail-ess

#> Warning: Bulk Effective Samples Size (ESS) is too low, indicating posterior means and medians may be unreliable.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#bulk-ess

#> Warning: Tail Effective Samples Size (ESS) is too low, indicating posterior variances and tail quantiles may be unreliable.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#tail-ess

#>                       Parameter Std_Median
#> 1                   (Intercept)  0.0000000
#> 3             Speciesversicolor -0.4221569
#> 5              Speciesvirginica  0.2440845
#> 2                   Petal.Width  0.7507258
#> 4 Speciesversicolor:Petal.Width  0.4730741
#> 6  Speciesvirginica:Petal.Width -0.2012074

# }