Parameters standardization

Compute standardized model parameters (coefficients).

standardize_parameters(
  model,
  method = "refit",
  ci = 0.95,
  robust = FALSE,
  two_sd = FALSE,
  verbose = TRUE,
  parameters,
  ...
)

standardize_posteriors(
  model,
  method = "refit",
  robust = FALSE,
  two_sd = FALSE,
  verbose = TRUE,
  ...
)

Arguments

model
A statistical model.

method
The method used for standardizing the parameters. Can be "refit" (default), "posthoc", "smart", "basic" or "pseudo". See 'Details'.

ci
Confidence Interval (CI) level

robust
Logical, if TRUE, centering is done by subtracting the median from the variables and dividing it by the median absolute deviation (MAD). If FALSE, variables are standardized by subtracting the mean and dividing it by the standard deviation (SD).

two_sd
If TRUE, the variables are scaled by two times the deviation (SD or MAD depending on robust). This method can be useful to obtain model coefficients of continuous parameters comparable to coefficients related to binary predictors, when applied to the predictors (not the outcome) (Gelman, 2008).

verbose
Toggle warnings on or off.

parameters
Deprecated.

...
For standardize_parameters(), arguments passed to parameters::model_parameters, such as:

ci_method, centrality for Bayesian models...

df_method for Mixed models ...

exponentiate, ...

etc.

Value

A data frame with the standardized parameters and their CIs.

Standardized parameters table.

Details

Methods:

refit: This method is based on a complete model re-fit with a standardized version of the data. Hence, this method is equal to standardizing the variables before fitting the model. It is the "purest" and the most accurate (Neter et al., 1989), but it is also the most computationally costly and long (especially for heavy models such as Bayesian models). This method is particularly recommended for complex models that include interactions or transformations (e.g., polynomial or spline terms). The robust (default to FALSE) argument enables a robust standardization of data, i.e., based on the median and MAD instead of the mean and SD. See standardize() for more details.

posthoc: Post-hoc standardization of the parameters, aiming at emulating the results obtained by "refit" without refitting the model. The coefficients are divided by the standard deviation (or MAD if robust) of the outcome (which becomes their expression 'unit'). Then, the coefficients related to numeric variables are additionally multiplied by the standard deviation (or MAD if robust) of the related terms, so that they correspond to changes of 1 SD of the predictor (e.g., "A change in 1 SD of x is related to a change of 0.24 of the SD of y). This does not apply to binary variables or factors, so the coefficients are still related to changes in levels. This method is not accurate and tend to give aberrant results when interactions are specified.

smart (Standardization of Model's parameters with Adjustment, Reconnaissance and Transformation - experimental): Similar to method = "posthoc" in that it does not involve model refitting. The difference is that the SD (or MAD if robust) of the response is computed on the relevant section of the data. For instance, if a factor with 3 levels A (the intercept), B and C is entered as a predictor, the effect corresponding to B vs. A will be scaled by the variance of the response at the intercept only. As a results, the coefficients for effects of factors are similar to a Glass' delta.

basic: This method is similar to method = "posthoc", but treats all variables as continuous: it also scales the coefficient by the standard deviation of model's matrix' parameter of factors levels (transformed to integers) or binary predictors. Although being inappropriate for these cases, this method is the one implemented by default in other software packages, such as lm.beta::lm.beta().

pseudo (for 2-level (G)LMMs only): In this (post-hoc) method, the response and the predictor are standardized based on the level of prediction (levels are detected with parameters::check_heterogeneity()): Predictors are standardized based on their SD at level of prediction (see also parameters::demean()); The outcome (in linear LMMs) is standardized based on a fitted random-intercept-model, where sqrt(random-intercept-variance) is used for level 2 predictors, and sqrt(residual-variance) is used for level 1 predictors (Hoffman 2015, page 342). A warning is given when a within-group varialbe is found to have access between-group variance.

Transformed Variables

When the model's formula contains transformations (e.g. y ~ exp(X)) method = "refit" might give different results compared to method = "basic" ("posthoc" and "smart" do not support such transformations): where "refit" standardizes the data prior to the transformation (e.g. equivalent to exp(scale(X))), the "basic" method standardizes the transformed data (e.g. equivalent to scale(exp(X))). See standardize() for more details on how different transformations are dealt with.

Generalized Linear Models

When standardizing coefficients of a generalized model (GLM, GLMM, etc), only the predictors are standardized, maintaining the interpretability of the coefficients (e.g., in a binomial model: the exponent of the standardized parameter is the OR of a change of 1 SD in the predictor, etc.)

References

Hoffman, L. (2015). Longitudinal analysis: Modeling within-person fluctuation and change. Routledge.

Neter, J., Wasserman, W., & Kutner, M. H. (1989). Applied linear regression models.

Gelman, A. (2008). Scaling regression inputs by dividing by two standard deviations. Statistics in medicine, 27(15), 2865-2873.

See also

Examples

library(effectsize)
data(iris)

model <- lm(Sepal.Length ~ Species * Petal.Width, data = iris)
standardize_parameters(model, method = "refit")

#> Parameter                     | Coefficient (std.) |        95% CI
#> ------------------------------------------------------------------
#> (Intercept)                   |               0.06 | [-1.43, 1.55]
#> Speciesversicolor             |              -0.17 | [-1.66, 1.33]
#> Speciesvirginica              |               0.19 | [-1.38, 1.76]
#> Petal.Width                   |               0.86 | [-0.32, 2.04]
#> Speciesversicolor:Petal.Width |               0.46 | [-0.88, 1.80]
#> Speciesvirginica:Petal.Width  |              -0.26 | [-1.52, 1.01]
#> 
#> # Standardization method: Refit
#> 

# \donttest{
standardize_parameters(model, method = "posthoc")

#> Parameter                     | Coefficient (std.) |        95% CI
#> ------------------------------------------------------------------
#> (Intercept)                   |               0.00 | [ 0.00, 0.00]
#> Speciesversicolor             |              -0.88 | [-2.07, 0.30]
#> Speciesvirginica              |               0.59 | [-0.69, 1.88]
#> Petal.Width                   |               0.86 | [-0.32, 2.04]
#> Speciesversicolor:Petal.Width |               0.46 | [-0.88, 1.80]
#> Speciesvirginica:Petal.Width  |              -0.26 | [-1.52, 1.01]
#> 
#> # Standardization method: Posthoc
#> 

standardize_parameters(model, method = "smart")

#> Parameter                     | Coefficient (std.) |        95% CI
#> ------------------------------------------------------------------
#> (Intercept)                   |               0.00 | [ 0.00, 0.00]
#> Speciesversicolor             |              -2.08 | [-4.85, 0.70]
#> Speciesvirginica              |               1.40 | [-1.62, 4.41]
#> Petal.Width                   |               0.86 | [-0.32, 2.04]
#> Speciesversicolor:Petal.Width |               1.07 | [-2.06, 4.19]
#> Speciesvirginica:Petal.Width  |              -0.60 | [-3.56, 2.35]
#> 
#> # Standardization method: Smart
#> 

standardize_parameters(model, method = "basic")

#> Parameter                     | Coefficient (std.) |        95% CI
#> ------------------------------------------------------------------
#> (Intercept)                   |               0.00 | [ 0.00, 0.00]
#> Speciesversicolor             |              -0.42 | [-0.98, 0.14]
#> Speciesvirginica              |               0.28 | [-0.33, 0.89]
#> Petal.Width                   |               0.86 | [-0.32, 2.04]
#> Speciesversicolor:Petal.Width |               0.38 | [-0.74, 1.50]
#> Speciesvirginica:Petal.Width  |              -0.33 | [-1.94, 1.28]
#> 
#> # Standardization method: Basic
#> 

# Robust and 2 SD
standardize_parameters(model, robust = TRUE)

#> Parameter                     | Coefficient (std.) |        95% CI
#> ------------------------------------------------------------------
#> (Intercept)                   |               0.18 | [-1.13, 1.49]
#> Speciesversicolor             |              -0.08 | [-1.40, 1.23]
#> Speciesvirginica              |               0.12 | [-1.24, 1.48]
#> Petal.Width                   |               0.93 | [-0.35, 2.21]
#> Speciesversicolor:Petal.Width |               0.50 | [-0.96, 1.95]
#> Speciesvirginica:Petal.Width  |              -0.28 | [-1.65, 1.09]
#> 
#> # Standardization method: Refit
#> 

standardize_parameters(model, two_sd = TRUE)

#> Parameter                     | Coefficient (std.) |        95% CI
#> ------------------------------------------------------------------
#> (Intercept)                   |               0.06 | [-1.43, 1.55]
#> Speciesversicolor             |              -0.17 | [-1.66, 1.33]
#> Speciesvirginica              |               0.19 | [-1.38, 1.76]
#> Petal.Width                   |               1.71 | [-0.65, 4.07]
#> Speciesversicolor:Petal.Width |               0.91 | [-1.76, 3.59]
#> Speciesvirginica:Petal.Width  |              -0.51 | [-3.04, 2.02]
#> 
#> # Standardization method: Refit
#> 

model <- glm(am ~ cyl * mpg, data = mtcars, family = "binomial")
standardize_parameters(model, method = "refit")

#> Parameter   | Coefficient (std.) |        95% CI
#> ------------------------------------------------
#> (Intercept) |              -0.58 | [-1.98, 0.70]
#> cyl         |               0.25 | [-1.54, 2.10]
#> mpg         |               2.10 | [-0.19, 5.28]
#> cyl:mpg     |              -0.36 | [-2.57, 1.54]
#> 
#> # Standardization method: Refit
#> 

standardize_parameters(model, method = "posthoc")

#> Parameter   | Coefficient (std.) |         95% CI
#> -------------------------------------------------
#> (Intercept) |               0.00 | [ 0.00,  0.00]
#> cyl         |               1.46 | [-4.63,  9.37]
#> mpg         |               3.36 | [-2.31, 12.59]
#> cyl:mpg     |              -0.20 | [-1.44,  0.86]
#> 
#> # Standardization method: Posthoc
#> 

standardize_parameters(model, method = "basic", exponentiate = TRUE)

#> Parameter   | Odds Ratio (std.) |            95% CI
#> ---------------------------------------------------
#> (Intercept) |              1.00 | [1.00,      1.00]
#> cyl         |              4.32 | [0.01,  11681.98]
#> mpg         |             28.80 | [0.10,  2.92e+05]
#> cyl:mpg     |              0.54 | [0.01,     13.94]
#> 
#> # Standardization method: Basic
#> 

# }

# \donttest{
if (require("lme4")) {
  m <- lmer(mpg ~ cyl + am + vs + (1|cyl), mtcars)
  standardize_parameters(m, method = "pseudo", df_method = "satterthwaite")
}

#> boundary (singular) fit: see ?isSingular

#> Warning: The following within-group terms have between-group variance:
#> 	am, vs
#> This can inflate standardized within-group parameters associated with
#> these terms. See help("demean", package = "parameters") for modeling
#> between- and within-subject effects.

#> Parameter   | Coefficient (std.) |         95% CI
#> -------------------------------------------------
#> (Intercept) |                    |               
#> cyl         |              -0.74 | [-1.23, -0.26]
#> am          |               0.47 | [ 0.01,  0.93]
#> vs          |               0.20 | [-0.47,  0.87]
#> 
#> # Standardization method: Pseudo
#> 

if (require("rstanarm")) {
  model <- stan_glm(Sepal.Length ~ Species + Petal.Width, data = iris, refresh = 0)
  # standardize_posteriors(model, method = "refit")
  # standardize_posteriors(model, method = "posthoc")
  # standardize_posteriors(model, method = "smart")
  head(standardize_posteriors(model, method = "basic"))
}

#>   (Intercept) Speciesversicolor Speciesvirginica Petal.Width
#> 1           0        0.02443670     -0.009429375   0.8437202
#> 2           0        0.04569722      0.139988415   0.6858118
#> 3           0        0.06528765      0.110950657   0.7460374
#> 4           0       -0.01040923     -0.030765243   0.8258345
#> 5           0       -0.01513676      0.044683846   0.7823281
#> 6           0        0.08562116      0.160969320   0.7319531

# }